+++ src/main/java/com/tomato/running/domain/user/service/GetMyInformationService.java
@@ -2,7 +2,7 @@ package com.tomato.running.domain.user.service;
 
 import com.tomato.running.domain.running.repository.RunningUserRepository;
 import com.tomato.running.domain.user.entity.User;
-import com.tomato.running.domain.user.controller.data.res.GetMyInformationResponseDto;
+import com.tomato.running.domain.user.presentation.data.res.GetMyInformationResponseDto;
 import com.tomato.running.domain.user.util.UserUtil;
 import com.tomato.running.global.annotation.ReadOnlyTransactionService;
